deployment/sandbox-v2/utils/minio_migration/requirements.txt
certifi==2020.12.5
minio==7.0.2
python-dotenv==0.15.0
urllib3==1.26.4
